XMLHttpRequest/tests/submissions/Opera/open-url-about-blank-window.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: open() resolving URLs (about:blank iframe)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <iframe src="about:blank"></iframe>
    <script>
      test(function() {
        var client = new self[0].XMLHttpRequest()
        client.open("GET", "folder.txt", false);
        client.send("");
        assert_equals(client.responseText, "top\n")
      })
    </script>
  </body>
</html>
